Remove 'deprecated' version 0.1.00 - See note
Hi Lloyd. The update you just merged for me was changing the version from 0.1.00 to 0.1.0 - which Lynn requested in preparation for balloting. I initially created a pull request that changed version value, and received an error from the pull check, In response, I manually added an entry deprecating version 0.1.00. (That's what you merged).

But I'm still getting a publisher error... because the entry deprecating the earlier 0.1.00 version isn't in the spec file generated by the publisher... which I probably should have predicted.  So this request removes the entry deprecating 0.1.00. 

Just let me know if there's a different thing I should do to get things back in sync. Sorry for the trouble. Thanks