Fix CI build URL and add version
The CI build doesn't recognize `main` as the default branch name, so I
updated the URL to manually include `/branch/main` so the links won't
be broken.

I also added 0.6.2 which is the version we will go to ballot with. Our
proposed ballot version was 0.6.0, and subsequent QA changes led to
incrementing the version number twice.